feat: improve homefragment, add responsive layouts
Now the HomeFragment presents general information about the podcast such as its logo, description and contact info. The episode logic will be moved to the SearchFragment. This fragment will be renamed accordingly.

Besides this, for the rest of fragments a landscape layout file has been added to ensure the correct view of all its elements. Giving to the user a responsive design.

This commit adds the new layouts, logic and modification for the work described above.