feat: change home and search fragment behavior
Now the HomeFragment only presents information about the Podcast.
On the other hand, SearchFragment has been renamed to EpisodeFragment,
this Fragment presents the episodes to the user. It also includes a
search option. This needs to be implemented.

Besides this, new drawable resources have been added and some have
been deleted.

This commit adds the new renamed files, logic, and updates the drawable
resources as explained above.