feat: first approach SearchFragment
The EpisodeFragment now has the option to search for episodes. This
episodes will appear in a new fragment named SearchFragment. For now
this episodes are the same as the one shown in EpisodeFragment.

We need to decide the approach to "search" episodes. This will be added
in following commits.

This commit adds the new files, changes, and logic to achieve the work
described above.