test: add test for Provider
Now the Provider class has testing for all its methods.

In those tests, we ensure that the data returned is actual data, and
not null data.

The idea is that, if empty data is provided to the UI (for example, no
internet during the execution of the app) then, the UI will be in charge
of notifying the user of this situation.

The Provider class, provides information obtained from the web, if there
is information, we are good. Otherwise, we return and empty object.