bug: wrong episode url
The url field in Episode had a wrong value (its description). This is fixed.

This also exposed that the unit tests must be improved. Until now we were checking only that data were there, but not the data itself. We need to ensure the data is a link (contains http(s)).

This commit add the fix for this bug.