feat: start work with reproducer
Currently, the layout associated to the media reproducer is updated
with the information of the clicked episode (from Episode or Search
fragment).

In order to do so, a new ViewModel, PlayerViewModel, has been created.
This ViewModel will contain also the logic associated to the reproducer.

This commit adds the new ViewModel class, along with the changes in
the different fragments to update the reproducer layout.