chore: continue improving layouts
The fragment_patreon.xml and fragment_social_network.xml layouts have
been modified to better fit the screens.

A new social network has been added, spotify.

A new library, TouchImageView, is used to allow to the user to zoom in
the news_fragment layout.

This commit adds the changes in the layouts, new spotify logo and the
use of the TouchImageView library.