bug: fix reproducer issue after rotating
The reproducer was playing audio even though there was no reproducer
layout enabled after rotating the screen.

To trigger this behavior you had to do the following:

* Remove the episode layout (swiping left) but NOT pausing the PlayerViewModel manually

* Rotate the screen. Then the episode would reproduce.

This was due to a corner case:

- When an episode is selected, we observe it in MainActivity (LiveData)
- When we remove the reproducer layout, the episode is still there.
Therefore, after a rotation (configuration change) it will be reproduced.

In order to fix this issue:

+ When the reproducer layout is removed, we "clean" the episode by
 setting all its fields to "".

+ When observing the episode, we check if its empty.

This commit adds the new logic to fix the issue, and new files necessary