chore: improve SocialNetworkFragment flow
The way the socialNetwork information is set has been modified. To do
so, a new model class (SocialNetwork.java) has been created.

The switch statement to set both, the imageView and
url, has been replaced for one that returns the imageView corresponding
 to the socialNetwork required.

Then we just need to set the onClickListener in that reference with the
url of the socialNetwork.

This commit adds the logic described above along with the new model.